feat: refactor SiteHeader component to share a single helper function (#180)
